re created TestBase, adding set up functions
