+++ CheckYourEligibility-Admin.Tests/CheckYourEligibility-Admin.Tests.csproj
@@ -32,6 +32,7 @@
   <ItemGroup>
     <ProjectReference Include="..\CheckYourEligibility-Admin\CheckYourEligibility-Admin.csproj" />
     <ProjectReference Include="..\CheckYourEligibility-FrontEnd.Services\CheckYourEligibility-FrontEnd.Services.csproj" />
+	<ProjectReference Include="..\CheckYourEligibility.TestBase\CheckYourEligibility.TestBase.csproj" />
   </ItemGroup>
 
   <ItemGroup>
