+++ CheckYourEligibility-Admin.Tests/Controllers/ApplicationControllerTests.cs
@@ -17,11 +17,12 @@ using CheckYourEligibility_FrontEnd.Models;
 using System.Security.Principal;
 using System.Security.Claims;
 using CheckYourEligibility_DfeSignIn.Models;
+using CheckYourEligibility.TestBase;
 
 namespace CheckYourEligibility_Parent.Tests.Controllers
 {
     [TestFixture]
-    public class ApplicationControllerTests
+    public class ApplicationControllerTests : TestBase
     {
         // mocks
         private ILogger<ApplicationController> _loggerMock;
@@ -40,66 +41,39 @@ namespace CheckYourEligibility_Parent.Tests.Controllers
         //system under test
         private ApplicationController _sut;
 
-        [SetUp]
-        public void SetUp()
+        public void TestSetUp()
         {
+            SetUp();
+        }
 
-            SetUpInitialMocks();
-            SetUpSessionData();
-            SetClaimsData();
-            SetUpHTTPContext();
-            // SetUpServiceMocks();
+        //[SetUp]
+        //public void SetUp()
+        //{
 
-            void SetUpInitialMocks()
-            {
-                _adminServiceMock = new Mock<IEcsServiceAdmin>();
-                _loggerMock = Mock.Of<ILogger<ApplicationController>>();
-                _sut = new ApplicationController(_loggerMock, _adminServiceMock.Object);
+        //    SetUpInitialMocks();
+        //    SetUpSessionData();
+        //    SetClaimsData();
+        //    SetUpHTTPContext();
+        //    // SetUpServiceMocks();
 
-            };
 
-            void SetUpHTTPContext()
-            {
-                _httpContext = new Mock<HttpContext>();
-                _httpContext.Setup(ctx => ctx.Session).Returns(_sessionMock.Object);
-                _httpContext.Setup(ctx => ctx.User).Returns(_userMock.Object);
-                _sut.ControllerContext.HttpContext = _httpContext.Object;
-            }
-
-            void SetClaimsData()
-            {
-                _userMock = new Mock<ClaimsPrincipal>();
-                var claimSchool = new Claim("organisation", Properties.Resources.ClaimSchool);
-                _userMock.Setup(x=>x.Claims).Returns(new List<Claim> { claimSchool,
+        void SetClaimsData()
+        {
+            _userMock = new Mock<ClaimsPrincipal>();
+            var claimSchool = new Claim("organisation", Properties.Resources.ClaimSchool);
+            _userMock.Setup(x => x.Claims).Returns(new List<Claim> { claimSchool,
                     new Claim($"http://schemas.xmlsoap.org/ws/2005/05/identity/claims/{ClaimConstants.NameIdentifier}", "123"),
                     new Claim($"http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress","test@test.com"),
                     new Claim($"http://schemas.xmlsoap.org/ws/2005/05/identity/claims/givenname","testFirstName"),
                     new Claim($"http://schemas.xmlsoap.org/ws/2005/05/identity/claims/surname","testSurname")
                 });
-            }
-
-            void SetUpSessionData()
-            {
-                _sessionMock = new Mock<ISession>();
-                var sessionStorage = new Dictionary<string, byte[]>();
-
-                _sessionMock.Setup(s => s.Set(It.IsAny<string>(), It.IsAny<byte[]>()))
-                                .Callback<string, byte[]>((key, value) => sessionStorage[key] = value);
-
-                _sessionMock.Setup(s => s.TryGetValue(It.IsAny<string>(), out It.Ref<byte[]>.IsAny))
-                            .Returns((string key, out byte[] value) =>
-                            {
-                                var result = sessionStorage.TryGetValue(key, out var storedValue);
-                                value = storedValue;
-                                return result;
-                            });
-            }
         }
+
 
         [TearDown]
         public void TearDown()
         {
-            _sut.Dispose();
+            TearDown();
         }
 
         [Test]
